fix(lazy-hydrate): optimized component loading
