+++ src/layouts/default.vue
@@ -9,6 +9,7 @@
 </template>
 
 <script>
+import { loadFonts } from '@/utils/fonts';
 import gpPageHeader from '@/components/page/Header';
 import {
   hydrateWhenVisible
@@ -79,60 +80,7 @@ export default {
 };
 
 if (process.client) {
-  var preloads = [];
-  if (navigator.connection && (navigator.connection.effectiveType === 'slow-2g' || navigator.connection.effectiveType === '2g')) {
-    preloads = document.querySelectorAll('link[rel=\'delay-prefetch\'][data-required=\'true\']');
-  } else {
-    preloads = document.querySelectorAll('link[rel=\'delay-prefetch\']');
-  }
-  prefetchFonts(Array.from(preloads));
-}
-
-function prefetchFonts (preloads, classList = []) {
-  let range = preloads.splice(0, Math.min(2, preloads.length));
-  global.requestIdleCallback(() => {
-    document.documentElement.classList.add(...classList.flat());
-    if (range.length) {
-      Promise.all(range.map((item) => {
-        return prefetchFont(item);
-      }))
-        .then((list) => {
-          prefetchFonts(preloads, list);
-          return;
-        })
-        .catch((e) => {
-          throw e;
-        });
-    }
-  });
-}
-
-function prefetchFont (item) {
-  return new Promise((resolve, reject) => {
-    item.onload = resolve;
-    item.onerror = reject;
-    item.rel = 'prefetch';
-  }).then((e) => {
-    return e.target.dataset;
-  }).then((options) => {
-    return loadFont(options);
-  })
-    .catch((e) => {
-      throw e;
-    });
-}
-
-function loadFont (options) {
-  return [
-    `font_${options.set}`,
-    `font_${options.set}_${options.weight}`
-  ];
-  // return document.fonts.load(`${options.width} 1em "${options.family}"`)
-  //   .then(() => {
-  //     return options.class;
-  //   }).catch((e) => {
-  //     throw e;
-  //   });
+  loadFonts();
 }
 
 </script>
